Unify Fp128 with Solinas-optimized arithmetic, delete SolinasFp128
Replace the generic Fp128<const MODULUS: u128> (binary-long-division via
U256) with the Solinas-optimized implementation. Fp128<const P: u128>
now uses [u64; 2] storage, fused schoolbook 2x2 + two-fold Solinas
reduction (~23 cycles/mul on AArch64/x86-64), and compile-time
validation that P = 2^128 - C with C < 2^64.

Delete SolinasFp128, SolinasParams, solinas128.rs, and u256.rs. All
call sites updated; prime type aliases (Prime128M13M4P0 etc.) are now
simple Fp128<...> aliases in fp128.rs. Blanket PseudoMersenneField impl
for all Fp128<P>.

Made-with: Cursor

+++ src/protocol/commitment/field_to_ring_reduction.rs
@@ -59,7 +59,7 @@ pub(crate) fn reduce_coeffs_to_ring_elements<F: FieldCore, const D: usize>(
     let mut out = Vec::with_capacity(outer_len);
     for i in 0..outer_len {
         let coeffs = std::array::from_fn(|j| {
-            let idx = i + ((j as usize) << outer_vars);
+            let idx = i + (j << outer_vars);
             coeffs[idx]
         });
         out.push(CyclotomicRing::from_coefficients(coeffs));
